level editor complete
actually was completed on 18.2.2020, but was too lazy to push.
since then, fixed a couple of bugs and added the boxfill brush. also, only all inactive layers are grayed out.

the boxfill brush makes the pencil brush kinda obsolete, since it's very easy to configure it to the radius of 0 (1 tile). whatever.